make binding parameters safe
