+++ src/statement/input.rs
@@ -1,4 +1,5 @@
 use {ffi, Return, Result, Raii, Handle, Statement};
+use super::types::FixedSizedType;
 
 /// Allows types to be used with `Statement::bind_parameter`
 pub unsafe trait InputParameter {
@@ -10,11 +11,39 @@ pub unsafe trait InputParameter {
 }
 
 impl<'a, 'b, S> Statement<'a, 'b, S> {
+    /// Binds a parameter to a parameter marker in an SQL statement.
+    ///
+    /// # Result
+    /// This method will destroy the statement and create a new one which may not outlive the bound
+    /// parameter. This is to ensure that the statement will not derefernce an invalid pointer
+    /// during execution.
+    ///
+    /// # Arguments
+    /// * `parameter_index` - Index of the marker to bind to the parameter. Starting at `1`
+    /// * `value` - Reference to bind to the marker
+    ///
+    /// # Example
+    /// ```
+    /// # use odbc::*;
+    /// # fn do_odbc_stuff() -> std::result::Result<(), Box<std::error::Error>> {
+    ///     let env = Environment::new()?.set_odbc_version_3()?;
+    ///     let conn = DataSource::with_parent(&env)?.connect("TestDataSource", "", "")?;
+    ///     let stmt = Statement::with_parent(&conn)?;
+    ///     let param = 1968;
+    ///     let stmt = stmt.bind_parameter(1, &param)?;
+    ///     let sql_text = "SELECT TITLE FROM MOVIES WHERE YEAR = ?";
+    ///     if let Data(mut stmt) = stmt.exec_direct(sql_text)? {
+    ///         // ...
+    ///     }
+    /// #   Ok(())
+    /// # }
+    /// ```
     pub fn bind_parameter<'c, T>(mut self,
                                  parameter_index: u16,
                                  value: &'c T)
                                  -> Result<Statement<'a, 'c, S>>
         where T: InputParameter,
+              T: ?Sized,
               'b: 'c
     {
         self.raii.bind_input_parameter(parameter_index, value).into_result(&self)?;
@@ -24,7 +53,8 @@ impl<'a, 'b, S> Statement<'a, 'b, S> {
 
 impl Raii<ffi::Stmt> {
     fn bind_input_parameter<T>(&mut self, parameter_index: u16, value: &T) -> Return<()>
-        where T: InputParameter
+        where T: InputParameter,
+              T: ?Sized
     {
         match unsafe {
                   ffi::SQLBindParameter(
@@ -48,7 +78,7 @@ impl Raii<ffi::Stmt> {
     }
 }
 
-unsafe impl<'a> InputParameter for &'a str {
+unsafe impl InputParameter for str {
     fn c_data_type(&self) -> ffi::SqlCDataType {
         ffi::SQL_C_CHAR
     }
@@ -92,9 +122,11 @@ unsafe impl InputParameter for String {
     }
 }
 
-unsafe impl InputParameter for u32 {
+unsafe impl<T> InputParameter for T
+    where T: FixedSizedType
+{
     fn c_data_type(&self) -> ffi::SqlCDataType {
-        ffi::SQL_C_ULONG
+        T::c_data_type()
     }
 
     fn column_size(&self) -> ffi::SQLULEN {
