+++ tests/input_parameter.rs
@@ -10,17 +10,14 @@ const B: &'static str = "SELECT B FROM TEST_TYPES WHERE B = ?;";
 const C: &'static str = "SELECT C FROM TEST_TYPES WHERE C = ?;";
 
 macro_rules! test_type {
-    ($t:ty, $c:expr, $e:expr) => ({
-        let param: $t = $e;
+    ($c:expr, $e:expr) => ({
 
         let env = Environment::new().unwrap().set_odbc_version_3().unwrap();
         let conn = DataSource::with_parent(&env).unwrap().connect("TestDataSource", "", "").unwrap();
         let stmt = Statement::with_parent(&conn).unwrap();
-        let stmt = stmt.bind_parameter(1, &param).unwrap();
+        let stmt = stmt.bind_parameter(1, $e).unwrap();
         if let Ok(Data(mut cursor)) = stmt.exec_direct($c){
-            if let Some(mut row) = cursor.fetch().unwrap(){
-                let value : $t = row.get_data(1).unwrap().unwrap();
-                assert_eq!(value, $e);
+            if let Some(_) = cursor.fetch().unwrap(){
             } else{
                 panic!("Result set has been empty");
             }
@@ -32,10 +29,72 @@ macro_rules! test_type {
 
 #[test]
 fn _ref_str() {
-    test_type!(&str, A, "Hello, World!")
+    let param = "Hello, World!";
+    test_type!(A, param)
 }
 
 #[test]
 fn _string() {
-    test_type!(String, A, String::from("Hello, World!"))
+    let param = String::from("Hello, World!");
+    test_type!(A, &param)
+}
+
+#[test]
+fn _i8() {
+    let param : i8 = 42;
+    test_type!(B, &param)
+}
+
+#[test]
+fn _u8() {
+    let param : u8 = 42;
+    test_type!(B, &param)
+}
+
+#[test]
+fn _i16() {
+    let param : i16 = 42;
+    test_type!(B, &param)
+}
+
+#[test]
+fn _u16() {
+    let param : u16 = 42;
+    test_type!(B, &param)
+}
+
+#[test]
+fn _i32() {
+    let param : i32 = 42;
+    test_type!(B, &param)
+}
+
+#[test]
+fn _u32() {
+    let param : u32 = 42;
+    test_type!(B, &param)
+}
+
+#[test]
+fn _i64() {
+    let param : i64 = 42;
+    test_type!(B, &param)
+}
+
+#[test]
+fn _u64() {
+    let param : u64 = 42;
+    test_type!(B, &param)
+}
+
+#[test]
+fn _f32() {
+    let param : f32 = 42.0;
+    test_type!(B, &param)
+}
+
+#[test]
+fn _f64() {
+    let param : f64 = 3.14;
+    test_type!(C, &param)
 }
